attempt to fix #2

+++ Models/DiceResult.cs
@@ -3,7 +3,7 @@ namespace TeleCasino.DiceGameService.Models;
 public class DiceResult
 {
     public required string Id { get; set; }
-    public int Wager { get; set; }
+    public decimal Wager { get; set; }
     public decimal Payout { get; set; }
     public decimal NetGain { get; set; }
     public required string VideoFile { get; set; }
